HFS+ : usage of signed integers and logging.

+++ fs/src/fs/org/jnode/fs/hfsplus/HfsPlusForkData.java
@@ -17,7 +17,7 @@
  * along with this library; If not, write to the Free Software Foundation, Inc., 
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
- 
+
 package org.jnode.fs.hfsplus;
 
 import java.io.IOException;
@@ -28,18 +28,24 @@ import org.jnode.util.BigEndian;
 public class HfsPlusForkData {
     public static final int FORK_DATA_LENGTH = 80;
     private static final int EXTENT_OFFSET = 16;
-    /** The size in bytes of the valid data in the fork. */
+    /**
+     * The size in bytes of the valid data in the fork.
+     */
     private long totalSize;
     /** */
-    private int clumpSize;
-    /** The total of allocation blocks use by the extents in the fork. */
-    private int totalBlock;
-    /** The first eight extent descriptors for the fork. */
+    private long clumpSize;
+    /**
+     * The total of allocation blocks use by the extents in the fork.
+     */
+    private long totalBlock;
+    /**
+     * The first eight extent descriptors for the fork.
+     */
     private ExtentDescriptor[] extents;
 
     /**
      * Create fork data from existing informations.
-     * 
+     *
      * @param src
      * @param offset
      */
@@ -47,20 +53,19 @@ public class HfsPlusForkData {
         byte[] data = new byte[FORK_DATA_LENGTH];
         System.arraycopy(src, offset, data, 0, FORK_DATA_LENGTH);
         totalSize = BigEndian.getInt64(data, 0);
-        clumpSize = BigEndian.getInt32(data, 8);
-        totalBlock = BigEndian.getInt32(data, 12);
+        clumpSize = BigEndian.getUInt32(data, 8);
+        totalBlock = BigEndian.getUInt32(data, 12);
         extents = new ExtentDescriptor[8];
         for (int i = 0; i < 8; i++) {
             extents[i] =
-                    new ExtentDescriptor(data, EXTENT_OFFSET +
-                            (i * ExtentDescriptor.EXTENT_DESCRIPTOR_LENGTH));
+                new ExtentDescriptor(data, EXTENT_OFFSET +
+                    (i * ExtentDescriptor.EXTENT_DESCRIPTOR_LENGTH));
         }
     }
 
     /**
-     * 
      * Create a new empty fork data object.
-     * 
+     *
      * @param totalSize
      * @param clumpSize
      * @param totalBlock
@@ -78,8 +83,8 @@ public class HfsPlusForkData {
     public byte[] write(byte[] dest, int destOffSet) {
         byte[] data = new byte[FORK_DATA_LENGTH];
         BigEndian.setInt64(data, 0, totalSize);
-        BigEndian.setInt32(data, 8, clumpSize);
-        BigEndian.setInt32(data, 12, totalBlock);
+        BigEndian.setInt32(data, 8, (int) clumpSize);
+        BigEndian.setInt32(data, 12, (int) totalBlock);
         for (int i = 0; i < extents.length; i++) {
             extents[i].write(data, EXTENT_OFFSET + (i * ExtentDescriptor.EXTENT_DESCRIPTOR_LENGTH));
         }
@@ -102,23 +107,24 @@ public class HfsPlusForkData {
         return totalSize;
     }
 
-    public int getClumpSize() {
+    public long getClumpSize() {
         return clumpSize;
     }
 
-    public int getTotalBlocks() {
+    public long getTotalBlocks() {
         return totalBlock;
     }
 
     public ExtentDescriptor getExtent(int index) {
         return extents[index];
     }
+
     /**
      * Read a block of data
      *
      * @param fileSystem the associated file system.
-     * @param offset the offset to read from.
-     * @param buffer the buffer to read into.
+     * @param offset     the offset to read from.
+     * @param buffer     the buffer to read into.
      * @throws java.io.IOException if an error occurs.
      */
     public void read(HfsPlusFileSystem fileSystem, long offset, ByteBuffer buffer) throws IOException {
@@ -140,7 +146,6 @@ public class HfsPlusForkData {
     }
 
     /**
-     * 
      * @param index
      * @param desc
      */
